+++ samples/custom-broadcast/src/main/java/com/appboy/custombroadcast/MainActivity.java
@@ -1,8 +1,8 @@
 package com.appboy.custombroadcast;
 
-import android.support.v7.app.AppCompatActivity;
 import android.content.Context;
 import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;