+++ samples/firebase-push/src/main/java/com/appboy/firebasepush/MainActivity.java
@@ -1,8 +1,8 @@
 package com.appboy.firebasepush;
 
 import android.content.Context;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;